+++ src/database/helpers/table_creation/all_table_creation_helpers.ts
@@ -1,0 +1,11 @@
+import { create_dynamodb_osu_score } from './create_dynamodb_osu_score';
+import { create_dynamodb_osu_beatmap } from './create_dynamodb_osu_beatmap';
+import { create_dynamodb_osu_beatmapset } from './create_dynamodb_osu_beatmapset';
+import { mods_to_bitwise } from './mods_to_bitwise';
+
+export {
+  create_dynamodb_osu_score,
+  create_dynamodb_osu_beatmap,
+  create_dynamodb_osu_beatmapset,
+  mods_to_bitwise,
+};
