+++ src/osu_api/ts_interfaces/Mods.ts
@@ -37,11 +37,3 @@ export enum Mods {
   ScoreV2 = 268435456,
   Mirror = 536870912,
 }
-
-export const mods_to_bitwise = (mods: Array<keyof typeof Mods>): number =>
-  mods.reduce((bitwise_accumulator, current_mod: keyof typeof Mods): number => {
-    if (typeof Mods[current_mod] !== 'number') {
-      throw new Error(`Mod - ${current_mod} - does not exist in enum`);
-    }
-    return bitwise_accumulator + Mods[current_mod];
-  }, 0);
